# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vaca EFA API
  version: 0.1.0
  description: API del sistema de control de vacaciones, saldos derivados y trazabilidad laboral.
servers:
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - cookieAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      security: []
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200': {description: Sesión creada}
        '401': {description: Credenciales inválidas}
  /api/v1/auth/refresh:
    post:
      security: []
      summary: Renovar cookies de sesión
      responses:
        '200': {description: Sesión renovada}
        '401': {description: Refresh inválido o expirado}
  /api/v1/auth/me:
    get:
      summary: Usuario de la sesión actual
      responses: {'200': {description: Usuario autenticado}, '401': {description: Sesión inválida}}
  /api/v1/auth/change-password:
    post:
      summary: Cambiar contraseña y revocar cookies actuales
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [currentPassword, newPassword], properties: {currentPassword: {type: string, format: password}, newPassword: {type: string, minLength: 8, format: password}}}}}
      responses: {'204': {description: Contraseña cambiada}}
...
  /api/v1/settlements:
    get:
      summary: Listar liquidaciones paginadas
      parameters:
        - {$ref: '#/components/parameters/Search'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: employmentId, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string, enum: [ACTIVE, ANULADA]}}
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
      responses: {'200': {description: Colección paginada en items}}
    post:
      summary: Registrar vacaciones disfrutadas o compensadas
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SettlementInput'}}}
      responses:
        '201': {description: Liquidación creada}
        '422': {$ref: '#/components/responses/BusinessRuleViolation'}
  /api/v1/settlements/{id}:
    patch:
      summary: Editar liquidación con control de versión
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SettlementInput'}}}
      responses:
        '200': {description: Liquidación actualizada}
        '409': {description: Conflicto de versión}
        '422': {$ref: '#/components/responses/BusinessRuleViolation'}
  /api/v1/vacation-settlements:
    get:
      summary: Listar liquidaciones paginadas (alias compatible)
      description: Alias de `/api/v1/settlements`; además de items devuelve data. Conserva filtros de empleo, búsqueda y estado.
      parameters:
        - {$ref: '#/components/parameters/Search'}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: employmentId, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string, enum: [ACTIVE, ANULADA]}}
      responses: {'200': {description: Colección paginada en items y data}}
    post:
      summary: Crear liquidación (alias compatible)
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SettlementInput'}}}
      responses:
        '201': {description: Liquidación creada}
        '422': {$ref: '#/components/responses/BusinessRuleViolation'}
  /api/v1/vacation-settlements/{id}:
    get:
      summary: Obtener liquidación y líneas originales (ruta compatible)
      parameters: [{$ref: '#/components/parameters/Id'}]
      responses:
        '200': {description: Liquidación completa}
        '404': {description: Liquidación no encontrada}
    patch:
      summary: Editar liquidación con control de versión (alias compatible)
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/SettlementInput'}}}
      responses:
        '200': {description: Liquidación actualizada}
        '409': {description: Conflicto de versión}
        '422': {$ref: '#/components/responses/BusinessRuleViolation'}
  /api/v1/vacation-settlements/{id}/annul:
    post:
      summary: Anular lógicamente una liquidación
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: {reason: {type: string, minLength: 3}}
      responses:
        '200': {description: Liquidación anulada}
        '409': {description: Conflicto de versión}
        '422': {$ref: '#/components/responses/BusinessRuleViolation'}
  /api/v1/vacation-settlements/import/preview:
    post:
      summary: Previsualizar importación CSV/XLSX de vacaciones disfrutadas
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/FileImportRequest'}}}
      responses:
        '200': {description: Lote PREVIEW, grupos, diferencias, conflictos y token}
        '409': {$ref: '#/components/responses/ImportConflict'}
        '422': {$ref: '#/components/responses/BusinessRuleViolation'}
  /api/v1/vacation-settlements/import/{batchId}/apply:
    post:
      summary: Aplicar importación autorizada de vacaciones disfrutadas
      description: "Registra las liquidaciones, crea los cronogramas COMPLETED correspondientes y ejecuta el barrido global de cierre con la fecha VACATION_CLOSURE_FROM_DATE: los períodos abiertos anteriores al corte que no estén protegidos por pendientes se cierran como 'Disfrutado (liquidación registrada)' cuando tienen liquidación o como 'Cerrado por migración' cuando no están contemplados. Las parcialidades quedan abiertas y se reportan en partiallyEnjoyedWarnings."
      parameters: [{$ref: '#/components/parameters/BatchId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/FileImportApplyRequest'}}}
      responses:
        '200':
          description: Lote aplicado transaccionalmente con conteos del barrido (closedByMigration, closedEnjoyedPeriods, partiallyEnjoyedWarnings)
        '409': {$ref: '#/components/responses/ImportConflict'}
        '422': {$ref: '#/components/responses/BusinessRuleViolation'}
  /api/v1/vacation-settlements/import/{batchId}:
    get:
      summary: Consultar lote de importación de vacaciones disfrutadas
      parameters: [{$ref: '#/components/parameters/BatchId'}]
      responses:
        '200': {description: Estado y trazabilidad del lote}
        '404': {$ref: '#/components/responses/ImportNotFound'}
...
  /api/v1/import/preview:
    post:
      summary: Validar y previsualizar un CSV de empleados (alias)
      description: Alias compatible de `/api/v1/worker-imports/preview`; ejecuta la misma validación y devuelve el mismo resultado sin la propiedad data.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EmploymentImportPreviewRequest'}
      responses:
        '200':
          description: Encabezados, filas mapeadas, filas normalizadas válidas, errores por fila y hash del payload autorizable
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EmploymentImportPreviewResponse'}
        '400': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: vaca_access
  responses:
    Unauthorized:
      description: Sesión ausente, inválida o expirada
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Forbidden:
      description: La sesión no tiene el permiso requerido
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    ValidationError:
      description: Cuerpo, query o límite de entrada inválido
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    ImportConflict:
      description: La clave o el lote pertenecen a otro contenido, están en proceso o no admiten la acción
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    BusinessRuleViolation:
      description: Una regla de negocio impide confirmar la operación
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    ImportNotFound:
      description: Lote de importación no encontrado
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
...
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string}}
    Search: {name: search, in: query, schema: {type: string}}
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 20}}
    IfMatch: {name: If-Match, in: header, schema: {type: integer, minimum: 1}}
    BatchId: {name: batchId, in: path, required: true, schema: {type: string}}
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string}
        password: {type: string, format: password}
...
    FileImportRequest:
      type: object
      required: [fileName]
      anyOf:
        - required: [content]
        - required: [contentBase64]
      properties:
        fileName: {type: string, minLength: 1}
        content: {type: string, description: Contenido CSV UTF-8}
        contentBase64: {type: string, description: Bytes CSV/XLSX codificados en Base64}
    FileImportApplyRequest:
      type: object
      required: [fileName, previewToken]
      anyOf:
        - required: [content]
        - required: [contentBase64]
      properties:
        fileName: {type: string, minLength: 1}
        content: {type: string, description: Contenido CSV UTF-8 idéntico al previsualizado}
        contentBase64: {type: string, description: Bytes CSV/XLSX idénticos a los previsualizados}
        previewToken: {type: string, minLength: 1}
    EmploymentImportPreviewRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          description: CSV UTF-8 completo, incluido el encabezado
    EmploymentImportRow:
      type: object
      description: Fila aún no validada. Puede usar nombres normalizados o encabezados admitidos por el importador; los errores se reportan por fila.
      properties:
        documentNumber: {type: string}
        fullName: {type: string}
        startDate: {type: string}
        endDate: {type: string}
        contractTypeName: {type: string}
        processName: {type: string}
        positionName: {type: string}
        supervisorName: {type: string}
      additionalProperties: true
    EmploymentImportPreviewResponse:
      type: object
      required: [headers, rows, validatedRows, totalRows, validRows, invalidRows, errors, payloadHash]
      properties:
        data:
          type: array
          description: Copia compatible de rows, presente únicamente en `/worker-imports/preview`.
          items: {$ref: '#/components/schemas/EmploymentImportRow'}
        headers:
          type: array
          items: {type: string}
        rows:
          type: array
          items: {$ref: '#/components/schemas/EmploymentImportRow'}
        validatedRows:
          type: array
          description: Resultado por fila. Solo las entradas valid=true incluyen data normalizada y pueden enviarse a la confirmación.
          items: {$ref: '#/components/schemas/EmploymentImportPreviewRow'}
        totalRows: {type: integer, minimum: 0}
        validRows: {type: integer, minimum: 0}
        invalidRows: {type: integer, minimum: 0}
        errors:
          type: array
          items: {$ref: '#/components/schemas/EmploymentImportError'}
        payloadHash: {type: string, pattern: '^[a-f0-9]{64}$', description: SHA-256 de las filas válidas normalizadas}
    EmploymentImportPreviewRow:
      type: object
      required: [row, valid, errors]
      properties:
        row: {type: integer, minimum: 2}
        valid: {type: boolean}
        data: {$ref: '#/components/schemas/EmploymentImportRow'}
        errors: {type: array, items: {type: string}}
...
    EmploymentImportError:
      type: object
      required: [row, message]
      properties:
        row: {type: integer, minimum: 0, description: Cero identifica un error general de persistencia}
        message: {type: string}
...
    SettlementInput:
      type: object
      required: [employmentId, enjoymentStartDate, enjoyedDays, compensatedDays, amountCOP, accountingDocument, allocations]
      properties:
        employmentId: {type: string}
        enjoymentStartDate: {type: string, format: date}
        enjoymentEndDate: {type: string, format: date}
        enjoyedDays: {type: integer, minimum: 0}
        compensatedDays: {type: integer, minimum: 0}
        amountCOP: {type: number, minimum: 0}
        accountingDocument: {type: string}
        allocations: {type: array, items: {type: object}}
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, BUSINESS_RULE_VIOLATION, INTERNAL_ERROR]
        message: {type: string}
        requestId: {type: string, description: Correlación de la petición; puede faltar en respuestas de compatibilidad directas}
        details:
          type: object
          additionalProperties: true
...